Files nach Java Packet Struktur neu organisiert test in serperate Datei ausgelagert
